- rollback last commit

+++ source/glest_game/menu/menu_background.cpp
@@ -160,8 +160,8 @@ void MenuBackground::createRainParticleSystem() {
 		rps->setSpeed(12.f/GameConstants::updateFps);
 		rps->setEmissionRate(25);
 		rps->setWind(-90.f, 4.f/GameConstants::updateFps);
-		rps->setPos(Vec3d(0.f, 25.f, 0.f));
-		rps->setColor(Vec4d(1.f, 1.f, 1.f, 0.2f));
+		rps->setPos(Vec3f(0.f, 25.f, 0.f));
+		rps->setColor(Vec4f(1.f, 1.f, 1.f, 0.2f));
 		rps->setRadius(30.f);
 
 		Renderer &renderer= Renderer::getInstance();
